# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Presidential Protocol & Events Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/profiles:
    get:
      operationId: listDashboardProfiles
      tags: [dashboard]
      summary: List dashboard profiles for an owner
      parameters:
        - name: ownerKey
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of dashboard profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DashboardProfile"
    post:
      operationId: createDashboardProfile
      tags: [dashboard]
      summary: Create a dashboard profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardProfileInput"
      responses:
        "201":
          description: Created dashboard profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardProfile"
  /dashboard/profiles/{id}:
    patch:
      operationId: updateDashboardProfile
      tags: [dashboard]
      summary: Update a dashboard profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: ownerKey
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DashboardProfileUpdate"
      responses:
        "200":
          description: Updated dashboard profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardProfile"
    delete:
      operationId: deleteDashboardProfile
      tags: [dashboard]
      summary: Delete a dashboard profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: ownerKey
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/search:
    get:
      operationId: searchContacts
      tags: [contacts]
      summary: Search contacts
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Matching contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
  /contacts/type/{type}:
    get:
      operationId: listContactsByType
      tags: [contacts]
      summary: List contacts by type
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contacts of a type
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
  /contacts/country/{countryCode}:
    get:
      operationId: listContactsByCountry
      tags: [contacts]
      summary: List contacts by country
      parameters:
        - name: countryCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contacts for a country
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get contact with notes, links and documents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    put:
      operationId: updateContact
      tags: [contacts]
      summary: Update contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contacts/{id}/notes:
    post:
      operationId: addContactNote
      tags: [contacts]
      summary: Add a note to a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactNoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactNote"
  /contacts/{id}/link-event:
    post:
      operationId: linkContactEvent
      tags: [contacts]
      summary: Link a contact to an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactEventLinkInput"
      responses:
        "201":
          description: Created link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactEventLink"
  /contacts/{id}/documents:
    post:
      operationId: addContactDocument
      tags: [contacts]
      summary: Attach a document to a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactDocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactDocument"
...
components:
  schemas:
    Contact:
      type: object
      required: [id, type, status, nameEn, pinned, confidential, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        status:
          type: string
        nameEn:
          type: string
        nameAr:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        nationality:
          type: ["string", "null"]
        countryCode:
          type: ["string", "null"]
        preferredLanguage:
          type: ["string", "null"]
        protocolTitle:
          type: ["string", "null"]
        protocolTitleAr:
          type: ["string", "null"]
        salutation:
          type: ["string", "null"]
        passportNumber:
          type: ["string", "null"]
        emiratesId:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        organization:
          type: ["string", "null"]
        organizationAr:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        jobTitleAr:
          type: ["string", "null"]
        roleInProtocol:
          type: ["string", "null"]
        classification:
          type: ["string", "null"]
        reportingLine:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        officeNumber:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        assistantContact:
          type: ["string", "null"]
        vipLevel:
          type: ["string", "null"]
        delegationRole:
          type: ["string", "null"]
        securityClearance:
          type: ["string", "null"]
        protocolRank:
          type: ["string", "null"]
        seatingPreference:
          type: ["string", "null"]
        giftPreference:
          type: ["string", "null"]
        specialRequirements:
          type: ["string", "null"]
        dietaryRequirements:
          type: ["string", "null"]
        accessibilityRequirements:
          type: ["string", "null"]
        culturalNotes:
          type: ["string", "null"]
        departmentKey:
          type: ["string", "null"]
        workflowRoles:
          type: ["string", "null"]
        permissionLevel:
          type: ["string", "null"]
        approvalAuthority:
          type: ["string", "null"]
        availability:
          type: ["string", "null"]
        taskCapacity:
          type: ["integer", "null"]
        activeTasks:
          type: ["integer", "null"]
        skills:
          type: ["string", "null"]
        extension:
          type: ["string", "null"]
        employeeType:
          type: ["string", "null"]
        eventResponsibilities:
          type: ["string", "null"]
        pinned:
          type: boolean
        confidential:
          type: boolean
        createdAt:
          type: string
        notes:
          type: array
          items:
            $ref: "#/components/schemas/ContactNote"
        eventLinks:
          type: array
          items:
            $ref: "#/components/schemas/ContactEventLink"
        documents:
          type: array
          items:
            $ref: "#/components/schemas/ContactDocument"
    ContactInput:
      type: object
      required: [type, nameEn]
      properties:
        type:
          type: string
        status:
          type: string
        nameEn:
          type: string
        nameAr:
          type: string
        gender:
          type: string
        nationality:
          type: string
        countryCode:
          type: string
        preferredLanguage:
          type: string
        protocolTitle:
          type: string
        protocolTitleAr:
          type: string
        salutation:
          type: string
        passportNumber:
          type: string
        emiratesId:
          type: string
        photoUrl:
          type: string
        organization:
          type: string
        organizationAr:
          type: string
        department:
          type: string
        jobTitle:
          type: string
        jobTitleAr:
          type: string
        roleInProtocol:
          type: string
        classification:
          type: string
        reportingLine:
          type: string
        email:
          type: string
        mobile:
          type: string
        officeNumber:
          type: string
        whatsapp:
          type: string
        assistantContact:
          type: string
        vipLevel:
          type: string
        delegationRole:
          type: string
        securityClearance:
          type: string
        protocolRank:
          type: string
        seatingPreference:
          type: string
        giftPreference:
          type: string
        specialRequirements:
          type: string
        dietaryRequirements:
          type: string
        accessibilityRequirements:
          type: string
        culturalNotes:
          type: string
        departmentKey:
          type: string
        workflowRoles:
          type: string
        permissionLevel:
          type: string
        approvalAuthority:
          type: string
        availability:
          type: string
        taskCapacity:
          type: integer
        activeTasks:
          type: integer
        skills:
          type: string
        extension:
          type: string
        employeeType:
          type: string
        eventResponsibilities:
          type: string
        pinned:
          type: boolean
        confidential:
          type: boolean
    ContactNote:
      type: object
      required: [id, contactId, body, createdAt]
      properties:
        id:
          type: integer
        contactId:
          type: integer
        body:
          type: string
        author:
          type: ["string", "null"]
        createdAt:
          type: string
    ContactNoteInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
        author:
          type: string
    ContactEventLink:
      type: object
      required: [id, contactId, eventId, createdAt]
      properties:
        id:
          type: integer
        contactId:
          type: integer
        eventId:
          type: integer
        role:
          type: ["string", "null"]
        rsvpStatus:
          type: ["string", "null"]
        createdAt:
          type: string
    ContactEventLinkInput:
      type: object
      required: [eventId]
      properties:
        eventId:
          type: integer
        role:
          type: string
        rsvpStatus:
          type: string
    ContactDocument:
      type: object
      required: [id, contactId, kind, label, createdAt]
      properties:
        id:
          type: integer
        contactId:
          type: integer
        kind:
          type: string
        label:
          type: string
        url:
          type: ["string", "null"]
        createdAt:
          type: string
    ContactDocumentInput:
      type: object
      required: [kind, label]
      properties:
        kind:
          type: string
        label:
          type: string
        url:
          type: string
...
    DashboardWidgetConfig:
      type: object
      required: [id, hidden]
      properties:
        id:
          type: string
        size:
          type: string
          enum: [sm, md, lg, full]
        hidden:
          type: boolean
        x:
          type: integer
        y:
          type: integer
        w:
          type: integer
        h:
          type: integer
    DashboardProfile:
      type: object
      required: [id, ownerKey, name, items, createdAt, updatedAt]
      properties:
        id:
          type: integer
        ownerKey:
          type: string
        name:
          type: string
        icon:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/DashboardWidgetConfig"
        createdAt:
          type: string
        updatedAt:
          type: string
    DashboardProfileInput:
      type: object
      required: [ownerKey, name, items]
      properties:
        ownerKey:
          type: string
        name:
          type: string
        icon:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/DashboardWidgetConfig"
    DashboardProfileUpdate:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/DashboardWidgetConfig"
...